SI, BP, AX and the DI copies are sixteen bits
An 8086 does 32-bit arithmetic in DX:AX and nowhere else, so a uint32_t
in this port is a claim that needs evidence. Auditing the 637 of them:
273 are C loop counters and 95 are flags, neither of which models a
register; two are the real thing, io_ticks reading the BIOS dword at
0040:006C and particle_height's idiv of DX:AX, whose comment already
explains why the product must not be folded into 32 bits.

That leaves the locals. This is the batch that is unambiguous because
the code names them after the register: 13 si, 6 bp, 5 ax and 13 d,
which is a working copy of di everywhere except hit_dirs - that one is a
direction and is a uint8_t. The two particle initialisers take a seed
that is AX and were declared for it.

Checked: morph_begin, morph_step, entity_paddle_fx, screen_level_done,
level_draw, level_intro, menu_particles_init, particle_init,
particle_random, ending_particles_init, ending_particles_tick and
ending_plot all agree with the original, and 2,000 frames on panel 8,
1,500 on level 11 and 1,200 through the bonus are identical throughout.

+++ reconstruct/src/game.h
@@ -1643,7 +1643,7 @@ void game_main(const char *dir, const char *levels);
  * this is where they were declared - nothing below is outstanding. The port's
  * one remaining stub is entity_unknown; see stubs.c.
  */
-void menu_particles_init(uint32_t ax_in);   /* 1ac2:5476 */
+void menu_particles_init(uint16_t ax_in);   /* 1ac2:5476 */
 void plot_pixel(uint16_t x, uint16_t y, uint32_t colour);
 void plot_pixel_xor(uint16_t x, uint16_t y, uint32_t colour);
 void menu_particles_tick(void);   /* 1ac2:53c2 */
@@ -1843,7 +1843,7 @@ void entities_clear(void);        /* 1ac2:055e */
 void level_between(void);         /* 1ac2:05f8 */
 void screen_game_over(void);      /* 1ac2:0473 */
 void ending_plot(uint16_t x, uint16_t y);     /* 1ac2:5add */
-void ending_particles_init(uint32_t ax); /* 1ac2:5a43 */
+void ending_particles_init(uint16_t ax); /* 1ac2:5a43 */
 void ending_particles_tick(void); /* 1ac2:5a56 */
 int32_t next_player(const char *dir);/* 1ac2:0d2e */
 void screen_results(const char *dir);   /* 1ac2:0ea3 */
